Added more towers

+++ src/enemies/regen_enemy.hpp
@@ -4,5 +4,5 @@
 
 class RegenEnemy : public Enemy {
 public:
-    RegenEnemy(const Vector4i& body, float x, float y);
+    RegenEnemy(const Vector4f& body, float x, float y);
 };
